+++ argilla-frontend/components/features/annotation/settings/Validation.vue
@@ -1,12 +1,12 @@
 <template>
-  <div :class="validations.length ? '--has-error' : null">
+  <div role="alert" :class="validations.length ? '--has-error' : null">
     <slot></slot>
     <template v-if="validations.length">
       <span
         class="--error-message"
         v-for="error in validations"
         :key="error"
-        v-html="error"
+        v-html="$t(error)"
       />
     </template>
   </div>