+++ argilla-frontend/nuxt.config.ts
@@ -53,31 +53,10 @@ const config: NuxtConfig = {
   },
 
   // Global CSS (https://go.nuxtjs.dev/config-css)
-  css: ["~assets/scss/base/base.scss"],
+  css: ["~assets/styles.scss"],
 
   // Plugins to run before rendering page (https://go.nuxtjs.dev/config-plugins)
-  plugins: [
-    { src: "~/plugins/logo" },
-
-    { src: "~/plugins/directives" },
-
-    { src: "~/plugins/di" },
-
-    { src: "~/plugins/language" },
-
-    { src: "~/plugins/plugins/axios.ts" },
-    { src: "~/plugins/plugins/axios-cache.ts" },
-    { src: "~/plugins/plugins/svg-icon.js" },
-    { src: "~/plugins/plugins/click-outside.js" },
-    { src: "~/plugins/plugins/toast.ts" },
-    { src: "~/plugins/plugins/copy-to-clipboard.js" },
-    { src: "~/plugins/plugins/filters.js" },
-    { src: "~/plugins/plugins/vue-draggable.js" },
-    { src: "~/plugins/plugins/platform.ts" },
-    { src: "~/plugins/plugins/language.ts" },
-    { src: "~/plugins/plugins/color-schema" },
-    { src: "~/plugins/plugins/color-generator.ts" },
-  ],
+  plugins: [{ src: "~/plugins" }],
 
   // Auto import components (https://go.nuxtjs.dev/config-components)
   components: [
@@ -94,7 +73,23 @@ const config: NuxtConfig = {
     // https://go.nuxtjs.dev/typescript
     "@nuxt/typescript-build",
     "@nuxtjs/composition-api/module",
-    ["@pinia/nuxt", { disableVuex: false }],
+    [
+      "@pinia/nuxt",
+      {
+        disableVuex: false,
+      },
+    ],
+    [
+      "nuxt-compress",
+      {
+        gzip: {
+          cache: true,
+        },
+        brotli: {
+          threshold: 10240,
+        },
+      },
+    ],
   ],
 
   // Modules (https://go.nuxtjs.dev/config-modules)
@@ -149,11 +144,12 @@ const config: NuxtConfig = {
     "/api/": {
       target: BASE_URL,
     },
+    "/share-your-progress": {
+      target: BASE_URL,
+    },
   },
-
   // Build Configuration (https://go.nuxtjs.dev/config-build)
   build: {
-    cssSourceMap: false,
     extend(config) {
       config.resolve.alias.vue = "vue/dist/vue.common";
       config.module.rules.push({
@@ -164,6 +160,26 @@ const config: NuxtConfig = {
         },
       });
     },
+    postcss: {
+      postcssOptions: {
+        order: "presetEnvAndCssnanoLast",
+        plugins: {
+          cssnano:
+            process.env.NODE_ENV === "production"
+              ? {
+                  preset: [
+                    "default",
+                    {
+                      discardComments: {
+                        removeAll: true,
+                      },
+                    },
+                  ],
+                }
+              : false,
+        },
+      },
+    },
     babel: {
       plugins: [["@babel/plugin-proposal-private-methods", { loose: true }]],
     },
@@ -173,6 +189,16 @@ const config: NuxtConfig = {
         keep_fnames: true,
       },
     },
+    extractCSS: true,
+    optimization: {
+      splitChunks: {
+        name: false,
+      },
+    },
+    filenames: {
+      css: ({ isDev }) => (isDev ? "[name].css" : "[contenthash].css"),
+    },
+    publicPath: "/_nuxt/",
   },
 
   // https://github.com/nuxt-community/style-resources-module