+++ argilla-frontend/v1/di/di.ts
@@ -29,7 +29,11 @@ import {
   JobRepository,
 } from "@/v1/infrastructure/repositories";
 
-import { useRole, useRoutes } from "@/v1/infrastructure/services";
+import {
+  useLocalStorage,
+  useRole,
+  useRoutes,
+} from "@/v1/infrastructure/services";
 import { useDataset } from "@/v1/infrastructure/storage/DatasetStorage";
 import { useRecords } from "@/v1/infrastructure/storage/RecordsStorage";
 import { useDatasets } from "@/v1/infrastructure/storage/DatasetsStorage";
@@ -63,13 +67,15 @@ import { OAuthLoginUseCase } from "@/v1/domain/usecases/oauth-login-use-case";
 import { GetEnvironmentUseCase } from "@/v1/domain/usecases/get-environment-use-case";
 import { GetWorkspacesUseCase } from "@/v1/domain/usecases/get-workspaces-use-case";
 import { GetDatasetQuestionsGroupedUseCase } from "@/v1/domain/usecases/get-dataset-questions-grouped-use-case";
+import { GetDatasetFieldsGroupedUseCase } from "@/v1/domain/usecases/get-dataset-fields-grouped-use-case";
 import { LoadUserUseCase } from "@/v1/domain/usecases/load-user-use-case";
 import { CreateDatasetUseCase } from "@/v1/domain/usecases/create-dataset-use-case";
 import { GetFirstRecordFromHub } from "@/v1/domain/usecases/get-first-record-from-hub";
+import { ExportDatasetToHubUseCase } from "@/v1/domain/usecases/export-dataset-to-hub-use-case";
 import { AuthLoginUseCase } from "@/v1/domain/usecases/auth-login-use-case";
 
 export const loadDependencyContainer = (context: Context) => {
-  const useAxios = useAxiosExtension(() => context.$axios);
+  const useAxios = useAxiosExtension(context);
   const useAuth = () => context.$auth;
 
   const dependencies = [
@@ -197,6 +203,10 @@ export const loadDependencyContainer = (context: Context) => {
       .withDependency(QuestionRepository)
       .build(),
 
+    register(GetDatasetFieldsGroupedUseCase)
+      .withDependency(FieldRepository)
+      .build(),
+
     register(GetDatasetSuggestionsAgentsUseCase)
       .withDependency(AgentRepository)
       .build(),
@@ -226,6 +236,10 @@ export const loadDependencyContainer = (context: Context) => {
       .build(),
 
     register(GetFirstRecordFromHub).withDependency(HubRepository).build(),
+
+    register(ExportDatasetToHubUseCase)
+      .withDependencies(DatasetRepository, useLocalStorage)
+      .build(),
   ];
 
   Container.register(dependencies);