+++ argilla-frontend/v1/domain/entities/dataset/Progress.ts
@@ -9,13 +9,19 @@ export class Progress {
   constructor(
     public readonly total: number,
     public readonly completed: number,
-    public readonly pending: number
+    public readonly pending: number,
+    public readonly users: Array<{ username: string }>
   ) {
     this.hasMetrics = total >= 0;
 
+    const percentagePending = (this.pending * 100) / this.total;
+    const percentageCompleted = (this.completed * 100) / this.total;
+
     this.percentage = {
-      pending: ((this.pending * 100) / this.total).toFixed(2),
-      completed: ((this.completed * 100) / this.total).toFixed(2),
+      pending: isNaN(percentagePending) ? "0.00" : percentagePending.toFixed(2),
+      completed: isNaN(percentageCompleted)
+        ? "0.00"
+        : percentageCompleted.toFixed(2),
     };
   }