+++ argilla-server/src/argilla_server/bulk/records_bulk.py
@@ -16,6 +16,7 @@ from datetime import datetime
 from typing import Dict, List, Sequence, Tuple, Union
 from uuid import UUID
 
+from datetime import UTC
 from fastapi.encoders import jsonable_encoder
 from sqlalchemy import select
 from sqlalchemy.ext.asyncio import AsyncSession
@@ -26,10 +27,11 @@ from argilla_server.api.schemas.v1.records_bulk import (
     RecordsBulk,
     RecordsBulkCreate,
     RecordsBulkUpsert,
-    RecordsBulkWithUpdateInfo,
+    RecordsBulkWithUpdatedItemIds,
 )
 from argilla_server.api.schemas.v1.responses import UserResponseCreate
 from argilla_server.api.schemas.v1.suggestions import SuggestionCreate
+from argilla_server.models.database import DatasetUser
 from argilla_server.webhooks.v1.enums import RecordEvent
 from argilla_server.webhooks.v1.records import notify_record_event as notify_record_event_v1
 from argilla_server.contexts import distribution
@@ -38,7 +40,7 @@ from argilla_server.contexts.records import (
     fetch_records_by_ids_as_dict,
 )
 from argilla_server.errors.future import UnprocessableEntityError
-from argilla_server.models import Dataset, Record, Response, Suggestion, Vector, VectorSettings
+from argilla_server.models import Dataset, Record, Response, Suggestion, Vector
 from argilla_server.search_engine import SearchEngine
 from argilla_server.validators.records import RecordsBulkCreateValidator, RecordUpsertValidator
 
@@ -109,12 +111,21 @@ class CreateRecordsBulk:
         self, records_and_responses: List[Tuple[Record, List[UserResponseCreate]]]
     ) -> List[Response]:
         upsert_many_responses = []
+        datasets_users = set()
         for idx, (record, responses) in enumerate(records_and_responses):
             for response_create in responses or []:
                 upsert_many_responses.append(dict(**response_create.model_dump(), record_id=record.id))
+                datasets_users.add((response_create.user_id, record.dataset_id))
 
         if not upsert_many_responses:
             return []
+
+        await DatasetUser.upsert_many(
+            self._db,
+            objects=[{"user_id": user_id, "dataset_id": dataset_id} for user_id, dataset_id in datasets_users],
+            constraints=[DatasetUser.user_id, DatasetUser.dataset_id],
+            autocommit=False,
+        )
 
         return await Response.upsert_many(
             self._db,
@@ -144,15 +155,11 @@ class CreateRecordsBulk:
             autocommit=False,
         )
 
-    @classmethod
-    def _metadata_is_set(cls, record_create: RecordCreate) -> bool:
-        return "metadata" in record_create.model_fields_set
-
 
 class UpsertRecordsBulk(CreateRecordsBulk):
     async def upsert_records_bulk(
         self, dataset: Dataset, bulk_upsert: RecordsBulkUpsert, raise_on_error: bool = True
-    ) -> RecordsBulkWithUpdateInfo:
+    ) -> RecordsBulkWithUpdatedItemIds:
         found_records = await self._fetch_existing_dataset_records(dataset, bulk_upsert.items)
 
         records = []
@@ -175,9 +182,14 @@ class UpsertRecordsBulk(CreateRecordsBulk):
                     external_id=record_upsert.external_id,
                     dataset_id=dataset.id,
                 )
-            elif self._metadata_is_set(record_upsert):
-                record.metadata_ = record_upsert.metadata
-                record.updated_at = datetime.utcnow()
+            else:
+                if record_upsert.is_set("metadata"):
+                    record.metadata_ = record_upsert.metadata
+                if record_upsert.is_set("fields"):
+                    record.fields = jsonable_encoder(record_upsert.fields)
+
+                if self._db.is_modified(record):
+                    record.updated_at = datetime.now(UTC)
 
             records.append(record)
 
@@ -193,7 +205,7 @@ class UpsertRecordsBulk(CreateRecordsBulk):
 
         await self._notify_upsert_record_events(records)
 
-        return RecordsBulkWithUpdateInfo(
+        return RecordsBulkWithUpdatedItemIds(
             items=records,
             updated_item_ids=[record.id for record in found_records.values()],
         )